backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
tortoise-orm==0.21.3
asyncpg==0.29.0
pydantic==2.7.2
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
httpx==0.27.0
pandas==2.2.2
python-dotenv==1.0.1
PyJWT==2.8.0
cryptography==42.0.7
